Remove spare board
Originally the idea was that update_tile would be functional, so we'd
avoid consing up a new board every time by having two boards and
flipping between them.  But once I realized that the collision rules
were going to be painful to resolve in parallel, I opted for the simpler
approach of preferring upper-leftmost moving bodies, and that means the
spare is no longer necessary.